requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
pyjwt
cachetools
google-generativeai
duckduckgo-search
